% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{get_diversification_rates_muhisse}
\alias{get_diversification_rates_muhisse}
\title{Extract estimated diversification rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)}
\usage{
get_diversification_rates_muhisse(model_fit, states)
}
\arguments{
\item{model_fit}{A muhisse fit object}

\item{states}{A vector of observed character states}
}
\value{
The estimated diversification rates formatted in a data.frame
}
\description{
A helper function that formats \code{hisse::MuHiSSE}'s diversification parameter estimates into a data frame
}
\examples{

library(hisse)
library(dplyr)

States <- c("mp", "mb", "fp", "fb")

# muhisse model
MH <- get(load("data/final.MuHiSSE2.Rsave"))
get_diversification_rates_muhisse(model_fit = MH, states=States)

# CID8 model
C8 <- get(load("data/final.CID8relax.Rsave"))
get_diversification_rates_muhisse(model_fit = C8, states=States)

# musse model, no hidden states
MU <- get(load("data/MuSSE.Rsave"))
get_diversification_rates_muhisse(model_fit = MU, states = States)
}
