% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{get_muhisse_rates}
\alias{get_muhisse_rates}
\title{Extract estimated transition and diversification rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)}
\usage{
get_muhisse_rates(model_fit, hidden_traits, character_states)
}
\value{
A list with two components, the transition rates formatted as a rate matrix and a data frame of diversification rates
}
\description{
This is a function that wraps \code{get_transition_matrix_muhisse} and \code{get_diversification_rates_muhisse}
so rates can be extracted in one step
}
\examples{

library(hisse)
library(dplyr)

States <- c("mp", "mb", "fp", "fb")

# muhisse model
MH <- get(load("data/final.MuHiSSE2.Rsave"))
get_muhisse_rates(model_fit = MH, hidden_traits=TRUE, character_states=States)
}
