% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{get_transitions_matrix_muhisse}
\alias{get_transitions_matrix_muhisse}
\title{Extract estimated transition rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)}
\usage{
get_transitions_matrix_muhisse(model_fit, hidden_states = TRUE, states)
}
\arguments{
\item{model_fit}{A muhisse fit object}

\item{hidden_states}{Logical whether the model contains hidden states}

\item{states}{A vector of state labels to replace hisse's 0/1 notation (see examples)}
}
\value{
The estimated transition rates formatted in a matrix
}
\description{
A helper function that formats \code{hisse::MuHiSSE}'s output into a matrix so transition rates between states can be easier to visualize.
}
\examples{

library(hisse)
library(dplyr)

# The translation should follow the order of states in model_fit$solution.
# In this example, hisse's 00, 01, 10, and 11 states correspond to
# marine-plankton, marine-benthos, freshwater-plankton, and freshwater-benthos

States <-
   expand.grid(observed=c("marine-plankton",
                          "marine-benthos",
                          "freshwater-plankton",
                          "freshwater-benthos"),
              hidden=LETTERS[1:2]) \%>\%
   mutate(Name=paste(observed, hidden, sep="_")) \%>\%
   select(Name) \%>\% unlist \%>\% unname

# muhisse model
MH <- get(load("data/final.MuHiSSE2.Rsave"))
get_transitions_matrix_muhisse(model_fit = MH, hidden_states = TRUE, states = States)

# CID8 model
C8 <- get(load("data/final.CID8relax.Rsave"))
# we have 8 hidden states, so the transition matrix is cumbersome
get_transitions_matrix_muhisse(model_fit = C8, hidden_states = TRUE, states = States)

# musse model, no hidden states
MU <- get(load("data/MuSSE.Rsave"))
get_transitions_matrix_muhisse(model_fit = MU, hidden_states = FALSE, states = c("mp", "mb", "fp", "fb"))

}
