% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{h_boxplot}
\alias{h_boxplot}
\title{Plot diversification rates estimated by a HiSSE model}
\usage{
h_boxplot(processed_hisse_recon, parameter = "turnover", x_label)
}
\arguments{
\item{processed_hisse_recon}{An object produced by \code{h_process_recon}}

\item{parameter}{parameter The diversification parameter to be plotted on the y axis. Possible options are turnover, extinct.frac, net.div, speciation, extinction}

\item{x_label}{Label for the x axis. This is usually the binary trait whose assosiation with diversification is being tested.}
}
\value{
A boxplot. Tip-associated rates (possibly model averaged) are plotted as points.
}
\description{
A function to plot a boxplot of (model-averaged) diversification rates in the alternative states. We can change the rate plotted on the y axis and modify the label for the x-axis (your binary 0/1 trait). Further modifications are straightforward with ggplot (see examples).
}
\examples{

library(hisse)
library(dplyr)
library(ggplot2)
library(viridis)
library(treeio)

asr <- get(load("data/hab.cid4.recon.Rsave"))
processed_hisse <- h_process_recon(hisse_recon=asr)
hisse_rates_plot <- h_boxplot(processed_hisse_recon=processed_hisse, parameter="turnover", x_label="habitat")

# change x axis tick labels
hisse_rates_plot +
 scale_x_discrete(breaks=c(0,1), labels=c("plankton", "benthos"))

# change the position of the legend
hisse_rates_plot +
 scale_x_discrete(breaks=c(0,1), labels=c("plankton", "benthos")) +
 theme(legend.position="top")

# use expressions (for greek letters) in the axis labels
hisse_rates_plot +
 scale_x_discrete(breaks=c(0,1), labels=c("plankton", "benthos")) +
 theme(legend.position="top") +
 labs(y=expression(paste(tau, "=", lambda, "+", mu))) +
 theme(axis.text.y=element_text(size=15)) +
 theme_grey()

}
