% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{h_process_recon}
\alias{h_process_recon}
\title{Prepare HiSSE marginal reconstruction for plotting}
\usage{
h_process_recon(hisse_recon)
}
\arguments{
\item{hisse_recon}{An object produced by \code{hisse::MarginRecon}, or a list of such objects over which we can model-average using \code{hisse::GetModelAveRates}}
}
\value{
A list with three components: \itemize{
 \item \code{tree_data} the phylogeny with associated data object with reconstructed ancestral, model-averaged node and tip states and rates
 \item \code{tip_rates/node_rates} a tibble of reconstructed ancestral, model-averaged tip/node states and rates.
 }
}
\description{
A function to process the output from \code{hisse::MarginRecon}. First we get model averaged rates for tips and nodes of the phylogeny. Then convert the tree from \code{phylo} format into a \code{treedata} class and assign the data.frame of mode-averaged rates to the \code{@data} slot. We output the node and tip rates tables and the reformated tree object for downstream plotting with \code{ggtree} and \code{ggplot}.
}
\examples{

asr <- get(load("data/cid4.recon.Rdata"))

processed_hisse <- h_process_recon(hisse_recon=asr)
processed_hisse$tip_rates
processed_hisse$node_rates
processed_hisse$tree_data

}
