% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{h_rate_recon}
\alias{h_rate_recon}
\title{Plot HiSSE model-averaged marginal ancestral state reconstruction for diversification rates}
\usage{
h_rate_recon(processed_recon, show_tip_labels = FALSE,
  parameter = "turnover", discrete = FALSE, breaks = seq(0, 1, 0.2),
  colors = c("yellow", "purple"), plot_as_waiting_time = FALSE,
  tree_layout = "rectangular", tree_direction = "right",
  time_axis_ticks = 10, open_angle = 10)
}
\arguments{
\item{processed_recon}{An object produced by \code{h_process_recon}}

\item{show_tip_labels}{Logical, whether to plot tip labels. Default is FALSE because it is difficult to plot legible tip labels for larger trees common in this type of analysis. See \code{?m_trait_recon} for a good manual solution.}

\item{parameter}{The diversification parameter to be mapped onto the tree. Possible options are turnover, extinct.frac, net.div, speciation, extinction}

\item{discrete}{Logical. Whether to discretize the distribution of reconstructed rates into bins}

\item{breaks}{A numeric vector of cut points for binning the rates. Passed internally to \code{cut}. The function checks whether 0 and max(rate) is in this vector and adds them if not. Scale appropriatelly if \code{plot_as_waiting_time=TRUE}.}

\item{colors}{Colors for the branches. If \code{discrete=TRUE} a vector of length matching the number of bins produced by the breaks. If \code{discrete=FALSE}, a vector of length two giving the min/max of a color gradient.}

\item{plot_as_waiting_time}{Whether to plot the rates (FALSE, default) or their inverse (waiting times)}

\item{tree_layout}{A layout for the tree. Available options are 'rectangular' (default), 'slanted', 'circular', 'fan' and 'radial'.}

\item{tree_direction}{'right' (default), 'left', 'up', or 'down' for rectangular and slanted tree layouts}

\item{time_axis_ticks}{numeric giving the number of ticks for the time axis (default=10)}

\item{open_angle}{The degrees of empty space between the first and last tip. Only works for \code{tree_layout = 'fan'} and allows for a little more space around axis tick labels.}
}
\value{
A plot of the phylogeny with branches colored by hisse-inferred marginal ancestral states.
}
\description{
A function to plot a (model-averaged) marginal ancestral reconstruction for the estimated diversification rates.
}
\examples{

data("diatoms")
processed_hisse <- h_process_recon(hisse_recon=diatoms$cid4_recon)

h_rate_recon(
   processed_recon = processed_hisse,
   parameter = "extinction", discrete=FALSE)

h_rate_recon(
   processed_recon = processed_hisse,
   parameter = "extinction", discrete=TRUE, breaks=c(0.3, 0.6, 1))

}
