% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{h_scatterplot}
\alias{h_scatterplot}
\title{Plot diversification rates estimated by a HiSSE model (HiSSE, BiSSE, CID2, CID4)}
\usage{
h_scatterplot(processed_hisse_recon, parameter = "turnover",
  x_label = "", state_names = c("0", "1"),
  plot_as_waiting_time = FALSE)
}
\arguments{
\item{processed_hisse_recon}{An object produced by \code{h_process_recon}}

\item{parameter}{The diversification parameter to be plotted on the y axis. Possible options are turnover, extinct.frac, net.div, speciation, extinction}

\item{x_label}{Label for the x axis. This is the binary trait whose assosiation with diversification is being tested.}

\item{state_names}{A character vector of length two giving the translation for states 0 and 1.}

\item{plot_as_waiting_time}{Logical. Whether to plot the rates or their inverse (waiting times)}
}
\value{
A jittered scatterplot of (model averaged) tip-associated rates.
}
\description{
A function to plot a scatterplot of (model-averaged) diversification rates in the alternative states. We can change the rate plotted on the y axis and modify the label for the x-axis (your binary 0/1 trait).
}
\examples{

asr <- get(load("data/cid4.recon.RData"))
processed_hisse <- h_process_recon(hisse_recon=asr)
hisse_rates_plot <- h_scatterplot(processed_hisse_recon=processed_hisse, parameter="turnover", x_label="habitat")

# modifications are easy with ggplot
# change x axis tick labels
hisse_rates_plot +
 scale_x_discrete(breaks=c(0,1), labels=c("plankton", "benthos"))

# change the position of the legend
hisse_rates_plot +
 scale_x_discrete(breaks=c(0,1), labels=c("plankton", "benthos")) +
 theme(legend.position="top")

# use expressions (for greek letters) in the axis labels
hisse_rates_plot +
 scale_x_discrete(breaks=c(0,1), labels=c("plankton", "benthos")) +
 theme(legend.position="top") +
 labs(y=expression(paste(tau, "=", lambda, "+", mu))) +
 theme(axis.text.y=element_text(size=15))

}
