% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{m_dotplot}
\alias{m_dotplot}
\title{Plot diversification rates estimated by a MuHiSSE model with means and standard deviations across tips}
\usage{
m_dotplot(processed_muhisse_recon, states_names = c("00", "01", "10",
  "11"), parameter = "turnover", colors = viridis(n = 4, option = 2,
  end = 0.7), plot_as_waiting_time = FALSE, bin_width = 0.1)
}
\arguments{
\item{processed_muhisse_recon}{An object produced by \code{m_process_recon}}

\item{parameter}{The diversification parameter to be plotted on the y axis. Possible options are turnover, extinct.frac, net.div, speciation, extinction}

\item{colors}{Colors for the lines and points}

\item{plot_as_waiting_time}{Whether to plot the rates (FALSE, default) or their inverse (waiting times)}

\item{bin_width}{The width of bins for the dotplot. Treat this as any histogram. Testing several different bin width values is recommended.}

\item{state_names}{Translation for the character states in the order 00, 01, 10, 11 (if wanted)}
}
\value{
A stacked dotplot of tip-associated rates (possibly model averaged) with summary statistics (mean +/- sd)
}
\description{
A function to plot a stacked dotplot of (model-averaged) diversification rates and summary statistics in the alternative states.
}
\examples{

data("diatoms")
processed_muhisse <- m_process_recon(muhisse_recon=diatoms$muhisse_recon)

m_dotplot(
 processed_muhisse_recon = processed_muhisse,
 parameter = "extinction",
 colors = viridis(n = 4, option=1, end=.8))
}
