% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{m_ridgelines}
\alias{m_ridgelines}
\title{Plot diversification rates estimated by a MuHiSSE model with means and standard deviations across tips}
\usage{
m_ridgelines(processed_muhisse_recon, states_names = c("00", "01", "10",
  "11"), parameter = "turnover", plot_as_waiting_time = FALSE,
  fill_colors = rep(NA, 4), line_colors = viridis(n = 4))
}
\arguments{
\item{processed_muhisse_recon}{An object produced by \code{m_process_recon}}

\item{parameter}{The diversification parameter to be plotted on the y axis. Possible options are turnover, extinct.frac, net.div, speciation, extinction}

\item{plot_as_waiting_time}{Whether to plot the rates (FALSE, default) or their inverse (waiting times)}

\item{fill_colors}{Colors for the density polygons}

\item{line_colors}{Colors for the lines and points}

\item{state_names}{Translation for the character states in the order 00, 01, 10, 11 (if wanted)}
}
\value{
A ridgeline plot of tip-associated rates (possibly model averaged), with carpet with the individual points and summary statistics (mean +/- sd)
}
\description{
A function to plot a ridgeline of (model-averaged) diversification rates in the alternative states.
}
\examples{

asr <- get(load("data/muhisse_relax_20_recon.Rdata"))
processed_muhisse <- m_process_recon(hisse_recon=asr)

m_ridgelines(
 processed_muhisse_recon = processed_muhisse,
 parameter = "extinction",
 line_colors = viridis(n = 4, option=1))
}
