% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{m_scatterplot}
\alias{m_scatterplot}
\title{Plot diversification rates estimated by a MuHiSSE model with means and standard deviations across tips}
\usage{
m_scatterplot(processed_recon, states_names = c("00", "01", "10", "11"),
  parameter = "turnover", colors = viridis(n = 4, option = 2, end =
  0.7), plot_as_waiting_time = FALSE)
}
\arguments{
\item{processed_recon}{An object produced by \code{m_process_recon}}

\item{states_names}{Translation for the character states in the order 00, 01, 10, 11 (if wanted)}

\item{parameter}{The diversification parameter to be plotted on the y axis. Possible options are turnover, extinct.frac, net.div, speciation, extinction}

\item{colors}{Colors for the lines and points}

\item{plot_as_waiting_time}{Whether to plot the rates (FALSE, default) or their inverse (waiting times)}
}
\value{
A jittered scatterplot of tip-associated rates (possibly model averaged) with summary statistics (mean +/- sd)
}
\description{
A function to plot a jittered scatterplot of (model-averaged) diversification rates and summary statistics in the alternative states.
}
\examples{

data("diatoms")
processed_muhisse <- m_process_recon(muhisse_recon=diatoms$muhisse_recon)

m_scatterplot(
 processed_recon = processed_muhisse,
 parameter = "extinction",
 colors = c("#000004FF", "#57157EFF", "#C43C75FF", "#FE9F6DFF"))

}
