% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{m_trait_recon_cp}
\alias{m_trait_recon_cp}
\title{Plot MuHiSSE model-averaged marginal ancestral state reconstruction for the trait with a two-dimensional colorplane}
\usage{
m_trait_recon_cp(processed_muhisse_recon, show_tip_label = FALSE,
  tree_layout = "rectangular", tree_direction = "right",
  time_axis_ticks = 10, open_angle = 10,
  focal_character = c("prob_0x", "prob_x0"), focal_character_label,
  second_character_label, colors = viridis(9)[c(5, 1, 9)])
}
\arguments{
\item{processed_muhisse_recon}{An object produced by \code{m_process_recon}}

\item{tree_layout}{A layout for the tree. Available options are 'rectangular' (default), 'slanted', 'circular', 'fan' and 'radial'.}

\item{tree_direction}{'right' (default), 'left', 'up', or 'down' for rectangular and slanted tree layouts}

\item{time_axis_ticks}{numeric giving the number of ticks for the time axis (default=10). Passed on to \code{pretty} internally, so the number of plotted ticks might not be exactly the same.}

\item{open_angle}{The degrees of empty space between the first and last tip. Only works for \code{tree_layout = 'fan'} and allows for a little more space around axis tick labels.}

\item{focal_character}{Specifies the x axis of the two-dimensional color key. Either \code{prob_0x} to plot the probability of state 0 for the first character, or \code{prob_x0} to plot the probability for state 0 for the second character.}

\item{focal_character_label}{Label for the x axis of the two-dimensional color gradient. This should match the focal probability.}

\item{second_character_label}{Label for the y axis of the two-dimensional color gradient.}

\item{colors}{A vector of three colors in the order: (1) zero color (color when the two traits are in state 0), (2) horizontal_color (color to interpolate towards state 1 of the focal character) and (2) vertical_color (color to interpolate towards state 1 of the second character). See \code{?colorplaner::color_projections} for details.}

\item{show_tip_labels}{Logical, whether to plot tip labels. Default is FALSE because it is difficult to plot legible tip labels for lerger trees common in this type of analysis. See \code{?m_trait_recon} for a good manual solution.}
}
\value{
A plot of the phylogeny with branches colored by muhisse-inferred marginal ancestral states.
}
\description{
A function to plot a MuHiSSE (model-averaged) marginal ancestral reconstruction for the trait data.
}
\examples{

data("diatoms")
processed_muhisse <- m_process_recon(muhisse_recon=diatoms$muhisse_recon)

m_trait_recon_cp(
 processed_muhisse_recon = processed_muhisse,
 tree_layout = "fan",
 focal_character = "prob_0x",
 focal_character_label = "p(marine)",
 second_character_label = "p(plankton)",
 colors = viridis(9)[c(5,1,9)]
 )

}
