% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{m_collect_rates}
\alias{m_collect_rates}
\title{Extract transition and diversification rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)}
\usage{
m_collect_rates(model_fit, hidden_traits, character_states)
}
\value{
A list with two components, the transition rates formatted as a rate matrix and a data frame of diversification rates
}
\description{
This is a function that wraps \code{m_transition_matrix} and \code{m_diversification_rates}
so rates can be extracted in one step
}
\examples{

States <- c("mp", "mb", "fp", "fb")

# muhisse model
MH <- get(load("data/final.MuHiSSE2.Rsave"))
m_collect_rates(model_fit = MH, hidden_traits=TRUE, character_states=States)
}
