% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilhisse.R
\name{m_diversification_rates}
\alias{m_diversification_rates}
\title{Extract estimated diversification rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)}
\usage{
m_diversification_rates(model_fit, states)
}
\arguments{
\item{model_fit}{A muhisse fit object}

\item{states}{A vector of observed character states}
}
\value{
The estimated diversification rates formatted in a data.frame
}
\description{
A helper function that formats \code{hisse::MuHiSSE}'s diversification parameter estimates into a data frame
}
\examples{

States <- c("mp", "mb", "fp", "fb")

# muhisse model
MH <- get(load("data/final.MuHiSSE2.RData"))
m_diversification_rates(model_fit = MH, states=States)

# CID8 model
C8 <- get(load("data/final.CID8relax.RData"))
m_diversification_rates(model_fit = C8, states=States)

# musse model, no hidden states
MU <- get(load("data/MuSSE.RData"))
m_diversification_rates(model_fit = MU, states = States)
}
